Conflicts:
	app/src/main/res/values/strings.xml

+++ app/src/main/java/union/union_vr1/Sqlite/DBAdapter_Temp_Canjes_Devoluciones.java
@@ -3,9 +3,16 @@ package union.union_vr1.Sqlite;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
+import android.database.CursorIndexOutOfBoundsException;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
+
+import java.text.DecimalFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.GregorianCalendar;
 
 import union.union_vr1.Conexion.DbHelper;
 
@@ -16,6 +23,7 @@ import union.union_vr1.Conexion.DbHelper;
 public class DBAdapter_Temp_Canjes_Devoluciones {
 
     public static final String temp_id_canjes_devoluciones = "_id";
+    public static final String temp_id_cabecera = "temp_id_cabecera";
     public static final String temp_id_comprob = "temp_in_id_comprob";
     public static final String temp_id_producto = "temp_in_id_producto";
     public static final String temp_nom_producto = "temp_te_nom_producto";
@@ -30,10 +38,16 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     public static final String temp_id_documento = "temp_id_documento";
     public static final String temp_id_motivo = "temp_id_motivo";
     public static final String temp_estado_producto = "temp_id_estado_producto";
+    public static final String temp_liquidacion = "temp_liquidacion";
+    public static final String temp_id_detalle = "temp_id_detalle";
+    public static final String temp_id_guia= "temp_id_guia";
+    public static final String temp_id_comprob_venta= "temp_id_comprob_venta";
+    public static final String temp_id_forma= "temp_id_forma";
+    public static final String temp_unidad = "temp_unidad";
 
     public static final String temp_estado = "temp_estado";
 
-
+    public DecimalFormat f = new DecimalFormat("##.00");
     public static final String TAG = "Temp_Venta_Detalle";
 
     private DbHelper mDbHelper;
@@ -46,20 +60,28 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     public static final String CREATE_TABLE_TEMP_CANJES_DEVOLUCIONES =
             "create table " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " ("
                     + temp_id_canjes_devoluciones + " integer primary key autoincrement,"
-                    + temp_id_comprob+" text,"
-                    + temp_id_producto+" text,"
-                    + temp_nom_producto+" text,"
-                    + temp_codigo_producto+" text,"
-                    + temp_cantidad+" text,"
-                    + temp_precio_unit+ " text,"
-                    + temp_importe+" text,"
-                    + temp_lote+" text,"
-                    + temp_fecha_emision+" text,"
-                    + temp_cliente+" text,"
-                    + temp_id_establecimiento+" text,"
-                    + temp_id_documento+" text,"
-                    + temp_id_motivo+" text,"
-                    + temp_estado_producto+" text,"
+                    + temp_id_cabecera + " text,"
+                    + temp_id_comprob + " text,"
+                    + temp_id_producto + " text,"
+                    + temp_nom_producto + " text,"
+                    + temp_codigo_producto + " text,"
+                    + temp_cantidad + " text,"
+                    + temp_precio_unit + " text,"
+                    + temp_importe + " text,"
+                    + temp_lote + " text,"
+                    + temp_fecha_emision + " text,"
+                    + temp_cliente + " text,"
+                    + temp_id_establecimiento + " text,"
+                    + temp_id_documento + " text,"
+                    + temp_id_motivo + " text,"
+                    + temp_estado_producto + " text,"
+                    + temp_id_detalle + " text,"
+                    + temp_id_guia + " text,"
+                    + temp_liquidacion + " text,"
+                    + temp_id_comprob_venta + " text,"
+                    + temp_id_forma + " text,"
+                    + temp_unidad + " text,"
+                    + Constants._SINCRONIZAR+ " text,"
                     + temp_estado + " integer);";
 
 
@@ -82,7 +104,7 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     }
 
     public long createTempCanjesDevoluciones(String id_comprob, String id_producto, String nom_producto, String codigo_producto, String cantidad, String precio_unit, String importe, String lote, String fecha_emision, String cliente,
-                                       String id_establecimiento, String id_documento, String id_motivo, String estado_producto, String estado) {
+                                             String id_establecimiento, String id_documento, String id_motivo, String estado_producto, String estado, String liquidacion, String idDetalle, String id_comprobante_venta,String id_forma, String unidad) {
 
         ContentValues initialValues = new ContentValues();
         initialValues.put(temp_id_comprob, id_comprob);
@@ -100,16 +122,165 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
         initialValues.put(temp_id_motivo, id_motivo);
         initialValues.put(temp_estado_producto, estado_producto);
         initialValues.put(temp_estado, estado);
+        initialValues.put(temp_liquidacion, liquidacion);
+        initialValues.put(temp_id_detalle, idDetalle);
+        initialValues.put(temp_fecha_emision, getDatePhone());
+        initialValues.put(temp_id_comprob_venta, id_comprobante_venta);
+        initialValues.put(temp_id_forma, id_forma);
+        initialValues.put(Constants._SINCRONIZAR,Constants._CREADO);
+        initialValues.put(temp_unidad, unidad);
 
         return mDb.insert(SQLITE_TABLE_Temp_Canjes_Devoluciones, null, initialValues);
     }
-    public Cursor listarCanjes(){
-        Cursor cursor = mDb.rawQuery("select * from "+SQLITE_TABLE_Temp_Canjes_Devoluciones+" where "+temp_id_motivo+"='2';",null);
-        return  cursor;
+
+    public void updateIdCabecera( String establec,String ESTADO) {
+        String id = "";
+        id = getIdForHeader(getDatePhone(), establec);
+        Cursor op = getAllOperacionByEstablec(establec,ESTADO);
+        if (op !=null) {
+
+            while (op.moveToNext()) {
+                Log.d("COUNTDETALLE",""+op.getCount());
+                ContentValues initialValues = new ContentValues();
+                initialValues.put(temp_id_cabecera, id);
+                mDb.update(SQLITE_TABLE_Temp_Canjes_Devoluciones, initialValues,
+                        temp_id_canjes_devoluciones + "=?", new String[]{op.getString(op.getColumnIndexOrThrow(temp_id_canjes_devoluciones))});
+
+            }
+        }
+
     }
-    public Cursor listarDevoluciones(){
-        Cursor cursor = mDb.rawQuery("select * from "+SQLITE_TABLE_Temp_Canjes_Devoluciones+" where "+temp_id_motivo+"='1';",null);
-        return  cursor;
+    public int updateIdDetalle( String id,int idGuia) {
+                ContentValues initialValues = new ContentValues();
+                initialValues.put(temp_id_guia, idGuia);
+                return mDb.update(SQLITE_TABLE_Temp_Canjes_Devoluciones, initialValues,
+                        temp_id_canjes_devoluciones + "=?", new String[]{id});
+    }
+    public int updateEstadoExport( String id) {
+        ContentValues initialValues = new ContentValues();
+        initialValues.put(Constants._SINCRONIZAR, Constants._EXPORTADO+"");
+        return mDb.update(SQLITE_TABLE_Temp_Canjes_Devoluciones, initialValues,
+                temp_id_canjes_devoluciones + "=?", new String[]{id});
+    }
+    public int updateEstadoExportAfter( String id) {
+        ContentValues initialValues = new ContentValues();
+        initialValues.put(Constants._SINCRONIZAR, Constants._ACTUALIZADO+"");
+        return mDb.update(SQLITE_TABLE_Temp_Canjes_Devoluciones, initialValues,
+                temp_id_canjes_devoluciones + "=?", new String[]{id});
+    }
+
+    public String getIdForHeader(String fecha, String establec) {
+        String id = "";
+        Cursor cursor = mDb.rawQuery("select MAX(" + temp_id_cabecera + ") as " + temp_id_cabecera + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_fecha_emision + "='" + fecha + "' and " + temp_id_establecimiento + "='" + establec + "'", null);
+        cursor.moveToFirst();
+        try {
+            String[] spl = cursor.getString(cursor.getColumnIndexOrThrow(temp_id_cabecera)).split("\\*");
+            int count = Integer.parseInt(spl[2]);
+            count = count + 1;
+            id = "" + fecha + "*" + establec + "*" + count;
+
+        } catch (NullPointerException e) {
+            id = "" + fecha + "*" + establec + "*" + 1;
+        }
+
+
+        return id;
+    }
+    public Cursor getAllOperacionEstablecimiento() {
+        Cursor cursor = mDb.rawQuery("select DISTINCT("+temp_id_establecimiento+") from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where "+Constants._SINCRONIZAR+"='"+Constants._ACTUALIZADO+"'", null);
+        return cursor;
+    }
+    public Cursor getAllOperacionByEstablec(String establec,String ESTADO) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_establecimiento + "='" + establec + "' and "+temp_fecha_emision+"='"+getDatePhone()+"' and "+Constants._SINCRONIZAR+"='"+ESTADO+"'", null);
+        return cursor;
+    }
+
+
+    public Cursor getAllOperacion(String idOperacion,String ESTADO) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_cabecera + "='" + idOperacion + "' and "+Constants._SINCRONIZAR+"='"+ESTADO+"' ", null);
+        return cursor;
+    }
+
+    public Cursor listarCanjes(String establec) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='2' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+Constants._CREADO+"';", null);
+        return cursor;
+    }
+
+    public String[] getInforHeader(String idEstablecimiento) {
+        Cursor cabecera;
+        cabecera = mDb.rawQuery("select * from m_evento_establec where ee_in_id_establec ='" + idEstablecimiento + "'", null);
+        String cliente = "", documento = "";
+        if (cabecera.moveToFirst()) {
+            cliente = cabecera.getString(cabecera.getColumnIndexOrThrow(DbAdaptert_Evento_Establec.EE_nom_cliente));
+            documento = cabecera.getString(cabecera.getColumnIndexOrThrow(DbAdaptert_Evento_Establec.EE_doc_cliente));
+        }
+        return new String[]{cliente, documento};
+    }
+
+    public Double[] getInfoCanje(String establec) {
+        Cursor cursor = mDb.rawQuery("select sum(" + temp_importe + ") as " + temp_importe + "," + temp_id_cabecera + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='2' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+Constants._CREADO+"';", null);
+        double total = 0.0000;
+        double igv = 0.0000;
+        double base = 0.0000;
+        cursor.moveToFirst();
+        if (cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe)) > 0) {
+            total = cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe));
+            base = total / 1.18;
+            igv = total - base;
+            Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base) + "IDCOMPROBANTE " + cursor.getString(cursor.getColumnIndexOrThrow(temp_id_cabecera)));
+        }
+        return new Double[]{Double.parseDouble(f.format(total)), Double.parseDouble(f.format(base)), Double.parseDouble(f.format(igv))};
+    }
+
+    public String[] getInfoCabeceraOperacion(String establec,String ESTADO) {
+        Cursor cursor = mDb.rawQuery("select sum(" + temp_importe + ") as " + temp_importe + "," + temp_id_cabecera + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+ESTADO+"';", null);
+        double total = 0.0000;
+        double igv = 0.0000;
+        double base = 0.0000;
+        String idCabecera = "";
+        cursor.moveToFirst();
+        if (cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe)) > 0) {
+            total = cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe));
+            base = total / 1.18;
+            igv = total - base;
+            idCabecera = cursor.getString(cursor.getColumnIndexOrThrow(temp_id_cabecera));
+            Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base) + "IDCOMPROBANTE " + cursor.getString(cursor.getColumnIndexOrThrow(temp_id_cabecera)));
+        }
+        return new String[]{Double.parseDouble(f.format(total)) + "", Double.parseDouble(f.format(base)) + "", Double.parseDouble(f.format(igv)) + "", idCabecera};
+    }
+
+    private String getDatePhone() {
+        Calendar cal = new GregorianCalendar();
+        Date date = cal.getTime();
+        SimpleDateFormat df = new SimpleDateFormat("dd/MM/yyyy");
+        String formatteDate = df.format(date);
+        return formatteDate;
+    }
+
+    public Cursor listarDevoluciones(String establec) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='1' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+Constants._CREADO+"';", null);
+        return cursor;
+    }
+    public Cursor listarDevolucionesImpresion(String establec) {
+        Cursor cursor = mDb.rawQuery("select distinct("+temp_id_guia+") as "+temp_id_canjes_devoluciones+","+temp_cliente+","+temp_id_forma+",sum("+temp_importe+") as "+temp_importe+"   from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='1' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+Constants._EXPORTADO+"';", null);
+        return cursor;
+    }
+
+    public Double[] getInfoDevoluciones(String establec) {
+        double total = 0.0000;
+        double igv = 0.00000;
+        double base = 0.00000;
+        Cursor cursor = mDb.rawQuery("select sum(" + temp_importe + ") as " + temp_importe + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='1' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "' and "+Constants._SINCRONIZAR+"='"+Constants._CREADO+"';", null);
+        cursor.moveToFirst();
+
+        if (cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe)) > 0) {
+            total = cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe));
+            base = total / 1.18;
+            igv = total - base;
+            Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base));
+        }
+        Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base));
+        return new Double[]{Double.parseDouble(f.format(total)), Double.parseDouble(f.format(base)), Double.parseDouble(f.format(igv))};
     }
 
     public void updateTempCanjesDevoluciones2(String idorig, String iddest, String vals) {
@@ -130,13 +301,20 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
 
 
     }
-    public int truncateCanjesDevoluciones(){
-        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones,null,null);
-    }
-    public int deleteCanjesDevoluciones(int id){
-        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, temp_id_canjes_devoluciones + " = ?", new String[] { id+"" });
+
+    public int truncateCanjesDevoluciones() {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + "", null);
+        if (cursor.moveToFirst()) {
+            return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, null, null);
+        } else {
+            return 1;
+        }
+
     }
 
+    public int deleteCanjesDevoluciones(int id) {
+        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, temp_id_canjes_devoluciones + " = ?", new String[]{id + ""});
+    }
 
 
 }
